Resolvido bug de Faces completamente transparentes. Era causada pelo fato de operações de divisão de inteiros ao invés de divisão de ponto flutuante

+++ Game/src/Face.cpp
@@ -13,8 +13,8 @@ Face::Face(float x, float y, Vec2 tileSize, TileMap *tileMap)
 	box.y= y;
 	sp.ScaleX(tileSize.x/sp.GetWidth());
 	sp.ScaleY(tileSize.y/sp.GetHeight());
-	sp.colorMultiplier = Color( 255*rand()/RAND_MAX, 255*rand()/RAND_MAX, 255*rand()/RAND_MAX );
-	sp.alpha = 255*rand()/RAND_MAX;
+	sp.colorMultiplier = Color( 255*(float)rand()/RAND_MAX, 255*(float)rand()/RAND_MAX, 255*(float)rand()/RAND_MAX );
+	sp.alpha = 255*(float)rand()/RAND_MAX;
 	box.w= sp.GetWidth();
 	box.h= sp.GetHeight();
 	components.emplace_back( new DragAndDrop(tileMap, true, true) );
